Move the TemplateParameters trait from ir::ty to ir::template
Organizationally, it makes more sense.
