+++ src/ir/context.rs
@@ -6,9 +6,9 @@ use super::item::{Item, ItemCanonicalPath, ItemSet};
 use super::item_kind::ItemKind;
 use super::module::{Module, ModuleKind};
 use super::named::{UsedTemplateParameters, analyze};
-use super::template::TemplateInstantiation;
+use super::template::{TemplateInstantiation, TemplateParameters};
 use super::traversal::{self, Edge, ItemTraversal};
-use super::ty::{FloatKind, TemplateParameters, Type, TypeKind};
+use super::ty::{FloatKind, Type, TypeKind};
 use BindgenOptions;
 use cexpr;
 use callbacks::ParseCallbacks;
