+++ src/ir/item.rs
@@ -8,9 +8,9 @@ use super::function::Function;
 use super::item_kind::ItemKind;
 use super::layout::Opaque;
 use super::module::Module;
-use super::template::AsNamed;
+use super::template::{AsNamed, TemplateParameters};
 use super::traversal::{EdgeKind, Trace, Tracer};
-use super::ty::{TemplateParameters, Type, TypeKind};
+use super::ty::{Type, TypeKind};
 use clang;
 use clang_sys;
 use parse::{ClangItemParser, ClangSubItemParser, ParseError, ParseResult};
